# Perf improvements in Clone, Sort, Merge and ReadCsv