--- a/src/Microsoft.Data/PrimitiveColumn.Sort.cs
+++ b/src/Microsoft.Data/PrimitiveColumn.Sort.cs
@@ -64,7 +64,6 @@ namespace Microsoft.Data
                     }
                     else
                         value = _columnContainer.Buffers[bufferIndex][index];
-
                 }
                 else
                 {