--- a/src/Microsoft.Data/PrimitiveColumn.cs
+++ b/src/Microsoft.Data/PrimitiveColumn.cs
@@ -324,7 +324,8 @@ namespace Microsoft.Data
             return ret;
         }
 
-        private PrimitiveColumn<T> CloneImplementation(BaseColumn mapIndices, Func<long, long?> getIndex, bool invertMapIndices = false)
+        private PrimitiveColumn<T> CloneImplementation<U>(PrimitiveColumn<U> mapIndices, Func<long, long?> getIndex, bool invertMapIndices = false)
+            where U : unmanaged
         {
             if (!mapIndices.IsNumericColumn())
                 throw new ArgumentException(String.Format(Strings.MismatchedValueType, Strings.NumericColumnType), nameof(mapIndices));
@@ -332,29 +333,82 @@ namespace Microsoft.Data
             if (mapIndices.Length > Length)
                 throw new ArgumentException(Strings.MapIndicesExceedsColumnLenth, nameof(mapIndices));
 
+            ReadOnlySpan<T> thisSpan = _columnContainer.Buffers[0].ReadOnlySpan;
+            long minRange = 0;
+            long maxRange = DataFrameBuffer<T>.MaxCapacity;
+            long maxCapacity = maxRange;
             PrimitiveColumn<T> ret = new PrimitiveColumn<T>(Name, mapIndices.Length);
             ret._columnContainer._modifyNullCountWhileIndexing = false;
-            if (invertMapIndices == false)
+            for (int b = 0; b < mapIndices._columnContainer.Buffers.Count; b++)
             {
-                for (long i = 0; i < mapIndices.Length; i++)
+                int index = b;
+                if (invertMapIndices)
+                    index = mapIndices._columnContainer.Buffers.Count - 1 - b;
+
+                ReadOnlyDataFrameBuffer<U> buffer = mapIndices._columnContainer.Buffers[index];
+                ReadOnlySpan<U> mapIndicesSpan = buffer.ReadOnlySpan;
+                ReadOnlySpan<long> mapIndicesLongSpan = default;
+                ReadOnlySpan<int> mapIndicesIntSpan = default;
+                Span<T> retSpan = DataFrameBuffer<T>.GetMutableBuffer(ret._columnContainer.Buffers[index]).Span;
+                if (mapIndices.DataType == typeof(long))
                 {
-                    T? value = _columnContainer[getIndex(i).Value];
-                    ret[i] = value;
-                    if (!value.HasValue)
+                    mapIndicesLongSpan = MemoryMarshal.Cast<U, long>(mapIndicesSpan);
+                }
+                if (mapIndices.DataType == typeof(int))
+                {
+                    mapIndicesIntSpan = MemoryMarshal.Cast<U, int>(mapIndicesSpan);
+                }
+                for (int i = 0; i < buffer.Length; i++)
+                {
+                    int spanIndex = i;
+                    if (invertMapIndices)
+                        spanIndex = buffer.Length - 1 - i;
+
+                    long rowIndex = mapIndicesIntSpan.IsEmpty ? mapIndicesLongSpan[spanIndex] : mapIndicesIntSpan[spanIndex];
+                    if (rowIndex < minRange || rowIndex >= maxRange)
+                    {
+                        int bufferIndex = (int)(rowIndex / maxCapacity);
+                        thisSpan = _columnContainer.Buffers[bufferIndex].ReadOnlySpan;
+                        minRange = bufferIndex * maxCapacity;
+                        maxRange = (bufferIndex + 1) * maxCapacity;
+                    }
+                    rowIndex -= minRange;
+                    T value = thisSpan[(int)rowIndex];
+                    bool isValid = IsValid(rowIndex);
+                    retSpan[i] = isValid ? value : default;
+                    if (!isValid)
+                    {
                         ret._columnContainer.NullCount++;
+                        ret._columnContainer.SetValidityBit(rowIndex, false);
+                    }
+                    else
+                    {
+                        ret._columnContainer.SetValidityBit(rowIndex, true);
+                    }
                 }
             }
-            else
-            {
-                long mapIndicesIndex = mapIndices.Length - 1;
-                for (long i = 0; i < mapIndices.Length; i++)
-                {
-                    T? value = _columnContainer[getIndex(mapIndicesIndex - i).Value];
-                    ret[i] = value;
-                    if (!value.HasValue)
-                        ret._columnContainer.NullCount++;
-                }
-            }
+            //if (invertMapIndices == false)
+            //{
+            //    for (long i = 0; i < mapIndices.Length; i++)
+            //    {
+            //        T? value = _columnContainer[getIndex(i).Value];
+            //        ret[i] = value;
+            //        if (!value.HasValue)
+            //            ret._columnContainer.NullCount++;
+            //    }
+
+            //}
+            //else
+            //{
+            //    long mapIndicesIndex = mapIndices.Length - 1;
+            //    for (long i = 0; i < mapIndices.Length; i++)
+            //    {
+            //        T? value = _columnContainer[getIndex(mapIndicesIndex - i).Value];
+            //        ret[i] = value;
+            //        if (!value.HasValue)
+            //            ret._columnContainer.NullCount++;
+            //    }
+            //}
             ret._columnContainer._modifyNullCountWhileIndexing = true;
             return ret;
         }