… tests/Microsoft.Data.DataFrame.Tests/DataFrame.IOTests.cs
… tests/Microsoft.Data.DataFrame.Tests/DataFrameTests.cs