Perf work towards dotnet/corefx#26845. 

Most of the changes are related to caching the Span<T> instead of randomly accessing Memory<T> every iteration. Some algorithms(for ex: Merge) need random access, so there's not much we can do there.