@pgovind 2019-09-05
Current Benchmark numbers:

          Method |        Mean |     Error |     StdDev |
---------------- |------------:|----------:|-----------:|
   ColumnIndexer |  7,848.6 us | 102.60 us |  90.953 us |
      ColumnSort | 24,445.9 us | 451.15 us | 422.007 us |
            Sort | 35,689.9 us | 306.84 us | 272.005 us |
            Clip |  5,140.9 us |  21.64 us |  19.187 us |
         GroupBy |    666.4 us |  10.62 us |   9.934 us |
 ColumnFillNulls |  4,726.6 us |  31.59 us |  28.002 us |
        LeftJoin | 10,519.7 us |  96.58 us |  85.615 us |
       RightJoin | 14,755.4 us | 212.28 us | 188.179 us |
       OuterJoin |  7,698.4 us |  77.77 us |  68.945 us |
       InnerJoin | 22,329.7 us | 432.96 us | 515.414 us |
       LeftMerge | 51,758.0 us | 455.56 us | 426.135 us |
      RightMerge | 37,633.9 us | 352.28 us | 329.522 us |
      OuterMerge | 55,329.4 us | 597.64 us | 499.054 us |
      InnerMerge | 44,772.9 us | 645.16 us | 571.917 us |


New Benchmark numbers with this patch:
          Method |        Mean |     Error |    StdDev |
---------------- |------------:|----------:|----------:|
   ColumnIndexer |  8,059.5 us | 159.74 us | 202.03 us |
      ColumnSort | 19,115.4 us | 379.31 us | 654.29 us |
            Sort | 20,351.7 us | 172.31 us | 161.18 us |
            Clip |  5,618.0 us |  70.38 us |  65.84 us |
         GroupBy |    751.9 us |  11.85 us |  11.09 us |
 ColumnFillNulls |  2,414.2 us |  24.26 us |  22.69 us |
        LeftJoin | 11,071.9 us |  75.09 us |  58.63 us |
       RightJoin |  6,078.1 us |  79.59 us |  74.45 us |
       OuterJoin |  7,822.8 us |  71.70 us |  59.87 us |
       InnerJoin |  5,012.9 us |  37.73 us |  35.29 us |
       LeftMerge | 24,607.5 us | 205.79 us | 182.42 us |
      RightMerge | 21,471.6 us | 383.92 us | 320.59 us |
      OuterMerge | 27,069.6 us | 536.02 us | 617.28 us |
      InnerMerge | 17,966.3 us | 296.75 us | 263.06 us |